+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This list is not intended to be all-encompassing - it will document major and br
 changes with their rationale when appropriate:
 
 ### v5.7.3.0 (uncut)
-- **http4k-*** : Upgrade some dependency versions
+- **http4k-*** : Upgrade some dependency versions, including Kotlin to 1.9.10
+- **http4k-htmx** : Added hyperscript.js webjar to the distribution.
 
 ### v5.7.2.0
 - **http4k-*** : Upgrade some dependency versions